book_management/requirements.txt
Django==5.1.4
djangorestframework==3.15.2
pymongo==4.10.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
Werkzeug==3.1.3
pytest==8.3.4
pytest-django==4.9.0
python-dotenv==1.0.1
whitenoise==6.8.2